Added Delay & Reverb effect configuration control
