+++ SwiftSimpler/Audio/Sample/Chain/EffectsConfiguration+Delay.swift
@@ -16,6 +16,7 @@ extension EffectsConfiguration {
         var feedback = Parameter(value: -30, range: -100...100)
         var time = Parameter(value: 0.05, range: 0...2)
         var mix = Parameter(value: 3, range: 0...100)
+        var bypass = false
     }
 }
