+++ SwiftSimpler/Views/EffectConfiguration/EffectViews/Controls/Knob/KnobSegment.swift
@@ -8,19 +8,25 @@
 import SwiftUI
 
 struct KnobSegment: View {
-    @Binding var parameter: Parameter
     let title: String
+
+    @Binding var parameter: Parameter
 
     var body: some View {
         GeometryReader { context in
+            
             VStack(spacing: 18) {
+                
                 Text(title)
                     .font(.headline)
                     .foregroundColor(Palette.white)
+                    .minimumScaleFactor(0.8)
                 
                 KnobControl(value: $parameter.value, range: parameter.range)
                     .frame(height: context.size.width)
+                
             }
         }
+        
     }
 }
