+++ SwiftSimpler/Views/EffectConfiguration/EffectViews/Controls/Knob/KnobValueFormatter.swift
