+++ SwiftSimpler/Views/EffectConfiguration/EffectViews/Controls/Knob/Shapes/KnobControlEdgesShape.swift
