+++ SwiftSimpler/Views/EffectConfiguration/EffectViews/Controls/Knob/Shapes/KnobControlPointerShape.swift
